chore: refactor package structure and update UserSecurityForm to use Java record
