+++ src/main/java/com/example/moabackend/domain/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.example.moabackend.domain.user.controller;
+
+public class UserController {
+}
