+++ src/main/java/com/example/moabackend/global/security/service/CustomUserDetailService.java
@@ -1,8 +1,8 @@
 package com.example.moabackend.global.security.service;
 
+import com.example.moabackend.domain.user.dto.UserSecurityForm;
 import com.example.moabackend.domain.user.exception.UserErrorCode;
-import com.example.moabackend.domain.user.persistence.repository.UserRepository;
-import com.example.moabackend.domain.user.persistence.repository.UserSecurityForm;
+import com.example.moabackend.domain.user.repository.UserRepository;
 import com.example.moabackend.global.exception.CustomException;
 import com.example.moabackend.global.security.info.UserPrincipal;
 import lombok.RequiredArgsConstructor;
